state-res: rewrite lexico-topo-sort fn
This fixes an issue where ties weren't broken correctly, which would
lead to the resolved state being completely wrong.

This version is also way easier to understand. I just threw the original
code away and started from scratch instead of trying to pinpoint where
exactly the bug was happening and fix just that part because it was very
confusing.

A lot of the tests now fail, but that's because the tests are wrong
too. The PDUs-under-test don't set `auth_events` and the mechanism
for generating `origin_server_ts` values can result in nonsense like
`m.room.create` being chronologically *after* other events. I have plans
for rewriting the test suite to deal with this.